talked to Marcus and made some major changes; see description
- rename most forein keys to "tablename_colname", some I kept for
reasons which will be clearer below.

- change "storage" table to location this will be a 1:N
- added an additional subtable called institution in case samples end up
at another place.
- reason why the name change is because I figure I can use this in
replace of "location_collection" in the sample table, so sample table
will have both storage and location of collection from the same table.

- created a ParentID table to keep track of lineage
- add row parentID FK in sample table

- created a logger table that logs all events done to the LIMS system.

- remove JSON field from sample and keep this in version table only.
- version JSON field will either store data associated with the sample
table/subject table or a record of the updated field.

- create junction table for subjects to consent (M:N) relationship

- created a diagnosis and diagnosis junction table for subject.

- also I found out from Stan that they can take biop from same subject
at different points of the diagnosis, to take this into consideration I
changed the forein key in sample from subject to a key referencing the
diagnosis junction table.

- created projects table and added this to fields of subject and sample

- issue of dates. All dates associated with data will be stored in JSON
field and modify date the database is stored as timestamp.

- add a link field to consent for in case we want to link to an actual
document
- changed all notes VARCHAR to TEXT fields instead
- change sex to ENUM type